Updated master to v2.0.3 of 3DGE. Does not include any work done in SDL2_migration or doom64-merge

+++ ddf/states.h
@@ -46,7 +46,7 @@ typedef struct state_s
 	short sprite;
 
     // frame ref (begins at 0.0)
-	float frame;
+	short frame;
 	
 	//number of frames for md5 (0 for individual frame)
 	short framerange;
